requirements.txt
crayons
pyyaml
requests
